+++ CHANGELOG.rst
@@ -18,6 +18,7 @@ Added
 
 - Add calculation of coefficient of determination for fits (R2) to quantify how well curve fits data:
   - Add ``HillCurve.r2`` attribute.
+  - Report coefficient of determination as "r2" in ``CurveFits.fitParams`` output.
 
 - Improve exception handling:
   - Add ``HillCurveFittingError``, and when fitting fails raise this exception rather than ``RuntimeError`` as was done previously.